+++ crates/ra_hir/src/code_model_impl.rs
@@ -1,4 +1,3 @@
-mod krate; // `crate` is invalid ident :(
 mod konst; // `const` is invalid ident :(
 mod module;
 pub(crate) mod function;
