--- crates/ra_hir/src/code_model_impl/krate.rs
@@ -1,22 +1,0 @@
-use crate::{
-    Crate, CrateDependency, AsName, Module, DefDatabase,
-};
-
-impl Crate {
-    pub(crate) fn dependencies_impl(&self, db: &impl DefDatabase) -> Vec<CrateDependency> {
-        let crate_graph = db.crate_graph();
-        crate_graph
-            .dependencies(self.crate_id)
-            .map(|dep| {
-                let krate = Crate { crate_id: dep.crate_id() };
-                let name = dep.as_name();
-                CrateDependency { krate, name }
-            })
-            .collect()
-    }
-    pub(crate) fn root_module_impl(&self, db: &impl DefDatabase) -> Option<Module> {
-        let module_id = db.crate_def_map(*self).root();
-        let module = Module { krate: *self, module_id };
-        Some(module)
-    }
-}
